+++ docs/sources/whatsnew.rst
@@ -8,6 +8,8 @@ Unreleased (Github version)
 User-visible Improvements
 ~~~~~~~~~~~~~~~~~~~~~~~~~
 
+- Add ``load_sample_list`` function to load samples in YAML format (see `YAML Samples <yaml-list.html>`_)
+
 - Added ``name()`` method to ``ReadSet`` objects, so you can do::